requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.2
Click==7.0
cryptography==2.6.1
Flask==1.0.2
future==0.17.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
pycparser==2.19
python-telegram-bot==11.1.0
six==1.12.0
Werkzeug==0.15.1
